requirements.txt
sqlalchemy-iris
langchain-iris
testcontainers-iris
llama-iris
sentence-transformers
langchain
fastembed 
openai 
tiktoken
python-dotenv
pandas
ipykernel